requirements.txt
tensorflow
pyspark
opencv-python
numpy
jupyter
pdf2image
pillow
pandas
scikit-learn
matplotlib

pyspark
tensorflow
jupyter
pandas
numpy
pdf2image
pillow
opencv-python
scikit-learn
matplotlib
seaborn